Modify demo.html date format attribute, add test and uncomment
There were typos: date-format needs to be data-date-format for the
dataToOptions function to work properly.

I also added a test case to support multiple formats. This applies when
typing values on the keyboard rather than clicking the calendar with the
mouse. Moment accepts arrays as format inputs. For display, the first
format in the array will be used.